+++ src/main/java/org/songeun/petdongne_server/address/application/service/AddressSearchService.java
@@ -1,19 +1,25 @@
-package org.songeun.petdongne_server.address.application;
+package org.songeun.petdongne_server.address.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
+import org.songeun.petdongne_server.address.application.dto.AddressBoundsSearchResponseDto;
+import org.songeun.petdongne_server.address.application.dto.AddressDtoMapper;
+import org.songeun.petdongne_server.address.application.dto.AddressSearchRequestDto;
 import org.songeun.petdongne_server.address.domain.AddressErrorStatus;
+import org.songeun.petdongne_server.address.domain.RegionAddressLevel;
+import org.songeun.petdongne_server.address.infrastructure.dto.LegalAddressBoundsSearchQueryResponseDto;
 import org.songeun.petdongne_server.address.infrastructure.dto.LegalAddressSearchQueryResponseDto;
 import org.songeun.petdongne_server.global.common.*;
 import org.songeun.petdongne_server.address.infrastructure.repository.LegalAddressSearchRepository;
 import org.songeun.petdongne_server.global.exception.BusinessException;
-import org.songeun.petdongne_server.global.search.KorSearchTextNormalizer;
-import org.songeun.petdongne_server.global.search.OrderedTokens;
-import org.songeun.petdongne_server.global.search.SynonymResolver;
-import org.songeun.petdongne_server.global.search.WhiteSpaceTokenizer;
+import org.songeun.petdongne_server.global.search.*;
+import org.songeun.petdongne_server.map.domain.ZoomLevel;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +28,7 @@ public class AddressSearchService {
     private final LegalAddressSearchRepository searchRepository;
     private final SynonymResolver regionSynonymResolver;
     private final WhiteSpaceTokenizer whiteSpaceTokenizer;
+    private final ZoomLevel zoomLevel;
 
     // todo: request dto 풀어주기
     public Slice<LegalAddressSearchQueryResponseDto> searchByText(AddressSearchRequestDto requestDto) {
